Create README.md
Kriti Raj is an absolute pro when it comes to crafting engaging content. With their exceptional skills in WordPress development and 2 years of experience in the blogging world, they know how to captivate readers. Their educational background in English writing and language adds an extra touch of finesse to their work. Kriti Raj is the go-to person for creating top-notch content and bringing websites to life! 🌟📝🔥